requirements.txt
Django==1.6.5
django-bootstrap3==4.7.0
django-extensions==1.3.8
django-bootstrap3-datetimepicker==2.2.3
psycopg2==2.5.3
